fix: update write_memory_to_file function signature to use size_t for data size

+++ include/cert_utils.h
@@ -23,6 +23,6 @@ SSL_CTX *create_client_ssl_context(void);
 
 /* File utility functions */
 char* read_file_to_memory(const char* filename, long* file_size);
-int write_memory_to_file(const char* filename, const char* data, long data_size);
+int write_memory_to_file(const char* filename, const char* data, size_t data_size);
 
 #endif /* CERT_UTILS_H */
